change int sizes on migrations

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,12 +20,12 @@ class CreateUsersTable extends Migration
             $table->date('join_date')->nullable();
             $table->string('discord_id')->unique();
             $table->string('discord_access_token')->nullable();
-            $table->bigInteger('avatar_resource_id')->unsigned()->nullable();
+            $table->integer('avatar_resource_id')->unsigned()->nullable();
             $table->date('date_of_birth');
             $table->smallInteger('country')->unsigned()->zerofill()->nullable()->default('Earth');
             $table->string('timezone', 9)->default('UTC+00:00');
             $table->json('data')->default(new Expression('(JSON_ARRAY())'));
-            $table->smallInteger('rank_id')->unsigned()->nullable();
+            $table->integer('rank_id')->unsigned()->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
